requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==3.0.10
Flask-Limiter==3.12
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
limits==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
pyotp==2.9.0
python-jwt==4.1.0
pytz==2025.2
qrcode==8.2
redis==6.4.0
requests==2.32.4
rich==13.9.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.0.3
wrapt==1.17.2